Change BuildTool::Place from u8 to u32 to match BT_* constants
- BuildTool::Place(u8) → Place(u32), eliminating casting friction
- BlockRegistry::get/name/is_wall now accept u32
- Replace remaining raw block type literals in apply_drag_shape and
  handle_build_placement with BT_* constants and bt_is!() macro
- Narrow to u8 only at make_block() call sites where grid data is written
- Remove redundant id32/bid aliases where id is already u32

+++ src/block_defs.rs
@@ -126,16 +126,16 @@ impl BlockRegistry {
         BlockRegistry { defs, wall_ids, placeable }
     }
 
-    pub fn get(&self, id: u8) -> Option<&BlockDef> {
+    pub fn get(&self, id: u32) -> Option<&BlockDef> {
         self.defs.get(id as usize).and_then(|d| d.as_ref())
     }
 
-    pub fn name(&self, id: u8) -> &str {
+    pub fn name(&self, id: u32) -> &str {
         self.get(id).map(|d| d.name.as_str()).unwrap_or("Unknown")
     }
 
-    pub fn is_wall(&self, id: u8) -> bool {
-        self.wall_ids.contains(&id)
+    pub fn is_wall(&self, id: u32) -> bool {
+        self.wall_ids.contains(&(id as u8))
     }
 
     pub fn tools_in_category<'a>(&'a self, cat: &str) -> Vec<&'a (u8, PlacementDef)> {
